Angry Birds prototype
One of my first projects where the user enters data for a target, then
is prompted to enter data for a projectile to hit the target. The
program responds to how the user did, and gives suggestions on how to
aim at the target for the next iteration of the game.